+++ src/main/java/com/example/BlogPost/entity/Reply.java
@@ -14,9 +14,9 @@ import java.time.LocalDateTime;
 public class Reply {
 
     @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer reply_id;
+    private Long reply_id;
 
-    private Integer author_id;
+    private Long author_id;
 
     @Column(nullable = false)
     private String contents;
